refactored into one class
